requirements.txt
PyYAML>=6.0
watchdog>=2.1.6
python-dotenv>=0.19.0
requests>=2.26.0
colorama>=0.4.4
python-json-logger>=2.0.2
